[rapids] all cuda ad rapids images are built successfully
custom_image_utils/shell_script_generator.py
* added retry function
* reduced debug output when -x is passed

examples/secure-boot/build-current-images.sh
* using the same timestamp for all images
* added perl code to disk_usage to extract maximum disk usage during installation
* making request for image and instance list once here instead of each
  time an image is about to be created

examples/secure-boot/install_gpu_driver.sh
* add is_debuntu to simplify common is_debian || is_ubuntu use case
* prepend clean to execute_with_retries to reduce maximum image size
* downloading assets to ram disk to reduce maximum image size
* added sync to key points of the script to allow accurate measurement
  of maximum disk usage
* moved initialization steps to new function prepare_to_install
* created exit_handler function which gets executed before script exits

examples/secure-boot/pre-init.screenrc
* re-ordered versions alphabetically

examples/secure-boot/pre-init.sh
* re-using cache of image and instance list from parent script
* increased machine type to accommodate ram disk
* updated disk image sizes based on disk usage analysis
* removed dask intermediate image

examples/secure-boot/rapids.sh
* removed spark-related logic (see spark-rapids)
* remove dask packages if they were installed
* moved initialization steps to new function prepare_to_install
* created exit_handler function which gets executed before script exits

+++ examples/secure-boot/pre-init.sh
@@ -17,9 +17,6 @@
 # pre-init.sh <dataproc version>
 
 set -e
-readonly timestamp="$(date +%F-%H-%M)"
-#readonly timestamp="2024-10-19-18-46"
-#readonly timestamp="2024-10-21-19-29"
 
 IMAGE_VERSION="$1"
 if [[ -z "${IMAGE_VERSION}" ]] ; then
@@ -62,25 +59,25 @@ fi
 function generate() {
   local extra_args="$*"
   local image_name="${PURPOSE}-${dataproc_version/\./-}-${timestamp}"
-  local img_list="$(gcloud compute images list --filter="name=( \"${image_name}\" )" 2>&1)"
-  set +e
-  echo ${img_list} | grep -q 'Listed 0 items.'
-  if [[ $? != 0 ]] ; then
-#  if gcloud compute images describe "${image_name}" > /dev/null ; then
+
+  local image="$(jq -r ".[] | select(.name == \"${image_name}\").name" "${tmpdir}/images.json")"
+
+  if [[ -n "${image}" ]] ; then
     echo "Image already exists"
     return
   fi
-  local instance_list="$(gcloud compute instances list --zones "${custom_image_zone}" --filter="name=( \"${image_name}-install\" )" 2>&1)"
-  echo ${instance_list} | grep -q 'Listed 0 items.'
-  if [[ $? != 0 ]]; then
-#  if gcloud compute instances describe "${image_name}-install" > /dev/null ; then
+
+  local instance="$(jq -r ".[] | select(.name == \"${image_name}-install\").name" "${tmpdir}/instances.json")"
+
+  if [[ -n "${instance}" ]]; then
     # if previous run ended without cleanup...
+    echo "cleaning up instance from previous run"
     gcloud -q compute instances delete "${image_name}-install" \
       --zone "${custom_image_zone}"
   fi
   set -xe
   python generate_custom_image.py \
-    --machine-type         "n1-standard-4" \
+    --machine-type         "n1-standard-8" \
     --accelerator          "type=nvidia-tesla-t4" \
     --image-name           "${image_name}" \
     --customization-script "${customization_script}" \
@@ -95,10 +92,7 @@ function generate() {
   set +x
 }
 
-function generate_from_dataproc_version() {
-  local dataproc_version="$1"
-  generate --dataproc-version "${dataproc_version}"
-}
+function generate_from_dataproc_version() { generate --dataproc-version "$1" ; }
 
 function generate_from_base_purpose() {
   generate --base-image-uri "https://www.googleapis.com/compute/v1/projects/${PROJECT_ID}/global/images/${1}-${dataproc_version/\./-}-${timestamp}"
@@ -106,15 +100,15 @@ function generate_from_base_purpose() {
 
 # base image -> cuda
 case "${dataproc_version}" in
-  "2.0-debian10" ) disk_size_gb="31" ;; # 40G   29G  9.1G  76% / # cuda-pre-init-2-0-debian10
-  "2.0-rocky8"   ) disk_size_gb="32" ;; # 33G   30G  3.4G  90% / # cuda-pre-init-2-0-rocky8
-  "2.0-ubuntu18" ) disk_size_gb="31" ;; # 36G   29G  7.0G  81% / # cuda-pre-init-2-0-ubuntu18
-  "2.1-debian11" ) disk_size_gb="37" ;; # 40G   35G  2.7G  93% / # cuda-pre-init-2-1-debian11
-  "2.1-rocky8"   ) disk_size_gb="35" ;; # 36G   33G  3.3G  91% / # cuda-pre-init-2-1-rocky8
-  "2.1-ubuntu20" ) disk_size_gb="36" ;; # 36G   34G  2.1G  95% / # cuda-pre-init-2-1-ubuntu20
-  "2.2-debian12" ) disk_size_gb="39" ;; # 40G   37G  1.1G  98% / # cuda-pre-init-2-2-debian12
-  "2.2-rocky9"   ) disk_size_gb="36" ;; # 54G   34G   20G  63% / # cuda-pre-init-2-2-rocky9
-  "2.2-ubuntu22" ) disk_size_gb="39" ;; # 39G   37G  2.4G  94% / # cuda-pre-init-2-2-ubuntu22
+  "2.0-debian10" ) disk_size_gb="38" ;; # 40G   31G  7.8G  80% / # cuda-pre-init-2-0-debian10
+  "2.0-rocky8"   ) disk_size_gb="35" ;; # 38G   32G  6.2G  84% / # cuda-pre-init-2-0-rocky8
+  "2.0-ubuntu18" ) disk_size_gb="37" ;; # 39G   30G  8.5G  79% / # cuda-pre-init-2-0-ubuntu18
+  "2.1-debian11" ) disk_size_gb="37" ;; # 39G   34G  4.1G  90% / # cuda-pre-init-2-1-debian11
+  "2.1-rocky8"   ) disk_size_gb="38" ;; # 41G   35G  6.1G  86% / # cuda-pre-init-2-1-rocky8
+  "2.1-ubuntu20" ) disk_size_gb="35" ;; # 37G   32G  4.4G  88% / # cuda-pre-init-2-1-ubuntu20
+  "2.2-debian12" ) disk_size_gb="38" ;; # 40G   35G  3.3G  92% / # cuda-pre-init-2-2-debian12
+  "2.2-rocky9"   ) disk_size_gb="39" ;; # 42G   36G  5.9G  86% / # cuda-pre-init-2-2-rocky9
+  "2.2-ubuntu22" ) disk_size_gb="38" ;; # 40G   35G  4.8G  88% / # cuda-pre-init-2-2-ubuntu22
 esac
 
 # Install GPU drivers + cuda on dataproc base image
@@ -122,40 +116,22 @@ PURPOSE="cuda-pre-init"
 customization_script="examples/secure-boot/install_gpu_driver.sh"
 time generate_from_dataproc_version "${dataproc_version}"
 
-# cuda image -> dask
+# cuda image -> rapids
 case "${dataproc_version}" in
-  "2.0-debian10" ) disk_size_gb="34" ;; # 40G   32G  6.4G  83% / # dask-pre-init-2-0-debian10
-  "2.0-rocky8"   ) disk_size_gb="35" ;; # 65G   33G   33G  51% / # dask-pre-init-2-0-rocky8
-  "2.0-ubuntu18" ) disk_size_gb="35" ;; # 36G   32G  4.3G  89% / # dask-pre-init-2-0-ubuntu18
-  "2.1-debian11" ) disk_size_gb="40" ;; # 40G   38G  111M 100% / # dask-pre-init-2-1-debian11
-  "2.1-rocky8"   ) disk_size_gb="37" ;; # 37G   36G  1.7G  96% / # dask-pre-init-2-1-rocky8
-  "2.1-ubuntu20" ) disk_size_gb="38" ;; # 41G   37G  4.3G  90% / # dask-pre-init-2-1-ubuntu20
-  "2.2-debian12" ) disk_size_gb="41" ;; # 46G   39G  4.4G  90% / # dask-pre-init-2-2-debian12
-  "2.2-rocky9"   ) disk_size_gb="39" ;; # 54G   37G   18G  68% / # dask-pre-init-2-2-rocky9
-  "2.2-ubuntu22" ) disk_size_gb="41" ;; # 44G   39G  4.9G  89% / # dask-pre-init-2-2-ubuntu22
+  "2.0-debian10" ) disk_size_gb="44" ;; # 47G   41G  4.0G  91% / # rapids-pre-init-2-0-debian10
+  "2.0-rocky8"   ) disk_size_gb="45" ;; # 49G   42G  7.0G  86% / # rapids-pre-init-2-0-rocky8
+  "2.0-ubuntu18" ) disk_size_gb="43" ;; # 45G   40G  4.9G  90% / # rapids-pre-init-2-0-ubuntu18
+  "2.1-debian11" ) disk_size_gb="46" ;; # 53G   43G  7.7G  85% / # rapids-pre-init-2-1-debian11
+  "2.1-rocky8"   ) disk_size_gb="48" ;; # 52G   45G  7.2G  87% / # rapids-pre-init-2-1-rocky8
+  "2.1-ubuntu20" ) disk_size_gb="45" ;; # 47G   42G  5.2G  89% / # rapids-pre-init-2-1-ubuntu20
+  "2.2-debian12" ) disk_size_gb="48" ;; # 51G   45G  3.8G  93% / # rapids-pre-init-2-2-debian12
+  "2.2-rocky9"   ) disk_size_gb="49" ;; # 54G   46G  8.2G  85% / # rapids-pre-init-2-2-rocky9
+  "2.2-ubuntu22" ) disk_size_gb="48" ;; # 52G   45G  7.7G  86% / # rapids-pre-init-2-2-ubuntu22
 esac
 
-# Install dask on cuda base image
-PURPOSE="dask-pre-init"
-customization_script="examples/secure-boot/dask.sh"
-time generate_from_base_purpose "cuda-pre-init"
-
-# dask image -> rapids
-case "${dataproc_version}" in
-   "2.0-debian10" ) disk_size_gb="52" ;; # 64G   49G   13G  80% / # rapids-pre-init-2-0-debian10
-   "2.0-rocky8"   ) disk_size_gb="52" ;; # 65G   50G   15G  78% / # rapids-pre-init-2-0-rocky8
-   "2.0-ubuntu18" ) disk_size_gb="53" ;; # 63G   49G   14G  78% / # rapids-pre-init-2-0-ubuntu18
-   "2.1-debian11" ) disk_size_gb="58" ;; # 64G   55G  6.2G  90% / # rapids-pre-init-2-1-debian11
-   "2.1-rocky8"   ) disk_size_gb="55" ;; # 65G   53G   12G  82% / # rapids-pre-init-2-1-rocky8
-   "2.1-ubuntu20" ) disk_size_gb="58" ;; # 63G   54G  9.1G  86% / # rapids-pre-init-2-1-ubuntu20
-   "2.2-debian12" ) disk_size_gb="59" ;; # 64G   56G  5.2G  92% / # rapids-pre-init-2-2-debian12
-   "2.2-rocky9"   ) disk_size_gb="56" ;; # 65G   54G   12G  83% / # rapids-pre-init-2-2-rocky9
-   "2.2-ubuntu22" ) disk_size_gb="60" ;; # 63G   56G  7.1G  89% / # rapids-pre-init-2-2-ubuntu22
-esac
-
-#disk_size_gb="65"
+#disk_size_gb="50"
 
 # Install rapids on dask base image
 PURPOSE="rapids-pre-init"
 customization_script="examples/secure-boot/rapids.sh"
-time generate_from_base_purpose "dask-pre-init"
+time generate_from_base_purpose "cuda-pre-init"
